Refactor wallet ingest logic
